Need to fix zero size elements still
And I think that's it. Although I discovered that:

* defining a char width on a wrap does nothing
* and that I probably need a vertical grid too.

So that a wrap can have grid or char heights and widths defined and the
layout is built based on that. I'm not sure how a mix of char and grid
dimensions should behave though. Maybe look at the flexbox grid spec for
tips?